fix: give the callbacks this binding provides real signature evidence (B2)
The prototype probe checks routes. For a callback parameter the managed side
declares nint it can only prove a pointer is passed, and what actually crosses
at run time is a managed function whose signature nothing had ever compared with
CNA's typedef. Two real defects were sitting in that gap.

NativeEventBridge.OnNativeEventWithSender took the sender as nint where CNA
passes a CNA_Handle, which is uint64_t. On a 64-bit target the two coincide,
which is exactly why nothing surfaced it; on a 32-bit target the sender is four
bytes short and every argument after it is read from the wrong place. The sender
is unused by the handler, so no test could have noticed.

CnaContentTypeReaderCallbacks.Destroy returned a result code where
CNA_ContentTypeReaderDestroyCallback returns void.

InteropCallbacks now declares, for each managed callback, a C function with the
signature derived from the managed member -- an UnmanagedCallersOnly method or a
function-pointer field of an interop struct -- and assigns it to CNA's typedef.
Nothing restates the header; the assignment is what fails. CALLBACKS_CHECKED=5,
CALLBACKS_UNRESOLVED=0, and an unresolvable pairing is a gate failure rather
than a silent skip.

Both defects were confirmed to be caught by reintroducing them. The sender one
produces "initialization of CNA_GraphicsResourceDisposingCallback from
incompatible pointer type void (*)(void*, void*)"; the destroy one no longer
compiles in C# at all, because the struct field and the method must agree.

The managed-callback-to-typedef pairing is listed rather than derived: a
function pointer is handed to a route at a call site, not declared as
implementing anything, so metadata cannot supply it. Five entries, each a
statement that the call site was read.

api-compat 257/257, 0 diagnostics. 225 + 560 + 140 tests. 9/9 negative controls
rejected, ABI policy passed.

+++ src/CNA.Framework/NativeEventBridge.cs
@@ -120,7 +120,7 @@ internal sealed class NativeEventBridge : IDisposable
         try
         {
             bridge._registration = subscribe(
-                (nint)(delegate* unmanaged[Cdecl]<nint, nint, void>)&OnNativeEventWithSender,
+                (nint)(delegate* unmanaged[Cdecl]<ulong, nint, void>)&OnNativeEventWithSender,
                 GCHandle.ToIntPtr(bridge._selfHandle));
         }
         catch
@@ -132,8 +132,17 @@ internal sealed class NativeEventBridge : IDisposable
         return bridge;
     }
 
+    /// <summary>
+    /// The sender is a <c>CNA_Handle</c>, which is a <c>uint64_t</c>, not a pointer.
+    ///
+    /// It was declared <c>nint</c> until B2's generated prototype probe compared the managed
+    /// callback with the C typedef. On a 64-bit target the two coincide and nothing is observably
+    /// wrong; on a 32-bit one <c>nint</c> is four bytes against an eight-byte argument, and every
+    /// parameter after it would be read from the wrong place. The sender is unused here, which is
+    /// why nothing had surfaced it.
+    /// </summary>
     [UnmanagedCallersOnly(CallConvs = new[] { typeof(CallConvCdecl) })]
-    private static void OnNativeEventWithSender(nint sender, nint context) => Dispatch(context);
+    private static void OnNativeEventWithSender(ulong sender, nint context) => Dispatch(context);
 
     [UnmanagedCallersOnly(CallConvs = new[] { typeof(CallConvCdecl) })]
     private static void OnNativeEvent(nint context) => Dispatch(context);
